requirements-dev.txt
aiofiles==23.2.1
aiohttp==3.9.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-cosmos==4.5.0
azure-identity==1.15.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
blinker==1.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.6
distro==1.9.0
firebase-admin==6.5.0
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.28.0
msal-extensions==1.1.0
msgpack==1.0.8
multidict==6.0.5
openai==1.6.1
packaging==24.0
portalocker==2.8.2
priority==2.0.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.0
Quart==0.19.4
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.24.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
langchain==0.2.5